OMAP3: MUSB: Fix multiple clock enable during init
MUSB initialization code for OMAP3 calls clk_enable/set_clock
in musb_platform_init() as well as in musb_platform_resume().
The latter is also called during init, so remove the call to
turn on the clocks from musb_platform_init().

Also there is no need to soft-reset the controller in
musb_platform_init() as the platform code has already done this.

Signed-off-by: Anand Gadiyar <gadiyar@ti.com>
Cc: Vikram Pandita <vikram.pandita@ti.com>