ASoC: dapm - add route checker for valid DSP paths
Add a DAPM API call to determine whether a DAPM audio path is valid.
i.e. it has a source and sink that are connected via kconrol mixer
and muxes.

This will be used by the DSP core to determine the runtime mappings
between FE and BE DAIs in order to run PCM operations.

Signed-off-by: Liam Girdwood <lrg@slimlogic.co.uk>