OMAP: iommu - disable interrupts if mmufault is not handled
If the mmufault callback registered by the user is not able to
manage the mmufault, we need to disable the interrupts to
avoid being interrupted again an again.

Before, when a mmu fault happended iommu_disable() was being called.
However, it gives an unhandle abort when user calls iommu_put because the
module clocks are already disabled.

Also, it removes irq acking from omap2_iommu_fault_isr function, due
to MMU tables are not updated with a valid pa, so there is no point
acking, the interrupt and status flag will be set again.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>