OMAP4: SMC: Allocate required 3MB mem for PA
This patch is to enable workspace allocation for SMC PA
to be operational . SMC requires 3M memory SDRAM on HS/EMU devices
for SDP <secure demand paging>.withim 1MB boundary alignment.

Out of this 3MB , 1920kb is for SMC workspace.

This is the reason the start physical addr is set 0xBFD00000.

Change-Id: I089525311d36e8a0417c21ddfc9e843f2f4ca48d
Signed-off-by: Trusted Logic <smc_support@trusted-logic.com>
Signed-off-by: Praneeth Bajjuri <praneeth@ti.com>