OMAP4: remoteproc: add rutime PM autosuspend support
Going to suspend and resuming for remote processor is a heavy
task and therefore much latency. For that reason we only going
to suspend after some inactivity time. By itself remoteproc
module can not know about activity on remoteproc (it is only
in charge of turning it on/off) that is why remoteproc add
new APIs to tell when it is used and then restart suspend
timeout.

Remote proc users needs to use that news apps. Besides if they
have issues when the remote processor goes to suspend without
their knowledge about that, they needs to register for suspend
notifications, returning an error in the notification callback
means they are using the remote processor and the autosuspend
is cancelled.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>