TILER: Select ARM_HAS_SG_CHAIN option
This patch enables the chaining of the scatter-gather lists,
to allow large number of pages to be passed into the TILER
driver for pinning. Without this support, pinning of
resolutions of even VGA fail.

This is being selected by TILER driver currently, so that
this is enabled for OMAP4 platforms. The ARM_HAS_SG_CHAIN
support patch is still not ratified for all ARM SoCs.

Change-Id: I3397be2acad3af23c44a1da2dba8e918e015a26c
Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Andy Gross <andy.gross@ti.com>