RPROC: add fatal error notifications for rproc users
Two new functions are added, so that rproc users can register/unregister
for fatal error notifications. Once a fatal error is notified the
remote processor is not usable anymore and rproc handle should be
release.

This patch includes MMU faults notifications, new fatal errors will
be added in the future.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>