Add API to update temperature thresholds
A temp sensor may have the ability to update it's temperature
thresholds so that it only reports changes that outside the
threshold window.

This API will be used to allow a governor to update the temperature
threshold window as the governor may not want every temperature
change unless it is outside the boundaries.

Signed-off-by: Dan Murphy <dmurphy@ti.com>