omap: rpres: skip the creation of already created devices
Some devices like:
- iva
- dsp
- fdif

are being created for pm purposes and create them again
is not recommended. For these devices the pm framework is
exporting api's to get them.

If a resource is already created use the api to get the
device and add the platform data needed for rpres to control
such device.

Change-Id: I6640883f911f4c21d204c20ded1ced104af5ed7c
Signed-off-by: Miguel Vadillo <vadillo@ti.com>