[WIP] TWL6040: skip setting the headset switch state
If the hardware doesn't truly support headset detection
then the switch should not be used to report the state of the
headset plug. Doing so confused the user space audio policy routing
decisions because the headset route is used at high priority and
caused other lower priority routes to not work.

Change-Id: Id6a66b9f0074f8528b40a3f64e21ae14c189f2de
Signed-off-by: Chris Kelly <c-kelly@ti.com>